+++ src/common/types/cast_helpers.cpp
@@ -209,7 +209,7 @@ std::string NumericHelper::ToString(uhugeint_t value) {
 }
 
 template <>
-int DecimalToString::DecimalLength(hugeint_t value, uint8_t width, uint8_t scale) {
+int DecimalToString::DecimalLength(hugeint_t value, uint32_t width, uint32_t scale) {
 	D_ASSERT(value > NumericLimits<hugeint_t>::Minimum());
 	int negative;
 
@@ -231,11 +231,11 @@ int DecimalToString::DecimalLength(hugeint_t value, uint8_t width, uint8_t scale
 	// integer length + 1 happens when the number is outside of that range
 	// in that case we print the integer number, but with one extra character ('.')
 	auto extra_numbers = width > scale ? 2 : 1;
-	return MaxValue(scale + extra_numbers, NumericHelper::UnsignedLength(value) + 1) + negative;
+	return MaxValue(static_cast<int>(scale) + extra_numbers, NumericHelper::UnsignedLength(value) + 1) + negative;
 }
 
 template <>
-string_t DecimalToString::Format(hugeint_t value, uint8_t width, uint8_t scale, Vector &vector) {
+string_t DecimalToString::Format(hugeint_t value, uint32_t width, uint32_t scale, Vector &vector) {
 	int length = DecimalLength(value, width, scale);
 	string_t result = StringVector::EmptyString(vector, NumericCast<idx_t>(length));
 
@@ -275,7 +275,7 @@ char *NumericHelper::FormatUnsigned(hugeint_t value, char *ptr) {
 }
 
 template <>
-void DecimalToString::FormatDecimal(hugeint_t value, uint8_t width, uint8_t scale, char *dst, idx_t len) {
+void DecimalToString::FormatDecimal(hugeint_t value, uint32_t width, uint32_t scale, char *dst, idx_t len) {
 	auto endptr = dst + len;
 
 	int negative = value.upper < 0;
