+++ src/common/types/decimal.cpp
@@ -5,26 +5,26 @@
 namespace duckdb {
 
 template <class SIGNED>
-string TemplatedDecimalToString(SIGNED value, uint8_t width, uint8_t scale) {
+string TemplatedDecimalToString(SIGNED value, uint32_t width, uint32_t scale) {
 	auto len = DecimalToString::DecimalLength<SIGNED>(value, width, scale);
 	auto data = make_unsafe_uniq_array_uninitialized<char>(UnsafeNumericCast<size_t>(len + 1));
 	DecimalToString::FormatDecimal<SIGNED>(value, width, scale, data.get(), UnsafeNumericCast<idx_t>(len));
 	return string(data.get(), UnsafeNumericCast<uint32_t>(len));
 }
 
-string Decimal::ToString(int16_t value, uint8_t width, uint8_t scale) {
+string Decimal::ToString(int16_t value, uint32_t width, uint32_t scale) {
 	return TemplatedDecimalToString<int16_t>(value, width, scale);
 }
 
-string Decimal::ToString(int32_t value, uint8_t width, uint8_t scale) {
+string Decimal::ToString(int32_t value, uint32_t width, uint32_t scale) {
 	return TemplatedDecimalToString<int32_t>(value, width, scale);
 }
 
-string Decimal::ToString(int64_t value, uint8_t width, uint8_t scale) {
+string Decimal::ToString(int64_t value, uint32_t width, uint32_t scale) {
 	return TemplatedDecimalToString<int64_t>(value, width, scale);
 }
 
-string Decimal::ToString(hugeint_t value, uint8_t width, uint8_t scale) {
+string Decimal::ToString(hugeint_t value, uint32_t width, uint32_t scale) {
 	auto len = DecimalToString::DecimalLength(value, width, scale);
 	auto data = make_unsafe_uniq_array_uninitialized<char>(UnsafeNumericCast<size_t>(len + 1));
 	DecimalToString::FormatDecimal(value, width, scale, data.get(), UnsafeNumericCast<idx_t>(len));
